complete all

+++ admin.php
@@ -3,6 +3,9 @@ include "vendor/autoload.php";
 
 use Libs\Database\MySQL;
 use Libs\Database\UsersTable;
+use Helpers\Auth;
+
+$auth = Auth::check();
 
 $table = new UsersTable(new MySQL);
 $users = $table->all();
@@ -26,6 +29,14 @@ $users = $table->all();
     <nav class="navbar bg-dark navbar-dark navbar-expand">
         <div class="container">
             <a href="#" class="navbar-brand">Admin</a>
+            <ul class="navbar-nav">
+                <li class="nav-item">
+                    <a href="profile.php" class="nav-link"><?= $auth->name ?></a>
+                </li>
+                <li class="nav-item">
+                    <a href="_actions/logout.php" class="nav-link">Logout</a>
+                </li>
+            </ul>
         </div>
     </nav>
 
@@ -64,13 +75,27 @@ $users = $table->all();
 
                     <td>
                         <div class="btn-group dropdown">
-                            <a href="#" class="btn btn-sm btn-outline-primary dropdown-toggle" data-bs-toggle="dropdown">Role</a>
-                            <div class="dropdown-menu dropdown-menu-dark">
-                                <a href="_actions/role.php?id=<?= $user->id ?>&role_id=1" class="dropdown-item">User</a>
-                                <a href="_actions/role.php?id=<?= $user->id ?>&role_id=2" class="dropdown-item">Manager</a>
-                                <a href="_actions/role.php?id=<?= $user->id ?>&role_id=3" class="dropdown-item">Admin</a>
-                            </div>
-                            <a href="_actions/delete.php?id=<?= $user->id ?> " class="btn btn-sm btn-outline-danger">Delete</a>
+                            <?php if ($auth->role_id == 3): ?>
+                                <a href="#" class="btn btn-sm btn-outline-primary dropdown-toggle" data-bs-toggle="dropdown">Role</a>
+                                <div class="dropdown-menu dropdown-menu-dark">
+                                    <a href="_actions/role.php?id=<?= $user->id ?>&role_id=1" class="dropdown-item">User</a>
+                                    <a href="_actions/role.php?id=<?= $user->id ?>&role_id=2" class="dropdown-item">Manager</a>
+                                    <a href="_actions/role.php?id=<?= $user->id ?>&role_id=3" class="dropdown-item">Admin</a>
+                                </div>
+                            <?php endif ?>
+
+                            <?php if ($auth->role_id >= 2): ?>
+                                <?php if ($user->suspended): ?>
+                                    <a href="_actions/unsuspend.php?id=<?= $user->id ?>" class="btn btn-sm btn-warning">Ban</a>
+
+                                <?php else: ?>
+                                    <a href="_actions/suspend.php?id=<?= $user->id ?>" class="btn btn-sm btn-outline-warning">Ban</a>
+                                <?php endif ?>
+
+                            <?php endif ?>
+                            <?php if ($auth->role_id == 3): ?>
+                                <a href="_actions/delete.php?id=<?= $user->id ?> " class="btn btn-sm btn-outline-danger">Delete</a>
+                            <?php endif ?>
                         </div>
                     </td>
                 </tr>
